+++ packages/form-render/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gopowerteam/form-render
 
+## 0.0.80
+
+### Patch Changes
+
+- Updated dependencies
+  - @gopowerteam/modal-render@0.0.44
+
 ## 0.0.79
 
 ### Patch Changes